feat: add Organic Growth Profiles and unify editor tooling
Introduce OrganicGenerationSettings as a ScriptableObject with a CreateAssetMenu and add sample Organic Growth Profile assets. Add asset/menu path constants for Organic Growth Profiles. Refactor the DungeonGenerator editor to support selecting or creating Organic Growth Profile assets, and replace older inline template UI with a richer editor: reorderable lists, template palette, default template mappings, chance/weight previews, and a TemplateEntry property drawer. Update runtime behavior to stop creating a default OrganicGenerationSettings instance on the generator, add null checks when generating, and support global template direction bias logic. Also adjust generation logging to be less severe for validation warnings and include various editor/whitespace fixes.

+++ DynamicDungeon_Tool/Assets/DynamicDungeon/ConstraintDungeon/Runtime/DungeonGenerationDiagnostics.cs
@@ -8,7 +8,7 @@ namespace DynamicDungeon.ConstraintDungeon
         private readonly Stopwatch stopwatch = new Stopwatch();
 
         public int attemptNumber;
-        public int seed;
+        public long seed;
         public int searchSteps;
         public int socketRejections;
         public int overlapRejections;
@@ -23,7 +23,7 @@ namespace DynamicDungeon.ConstraintDungeon
 
         public long ElapsedMilliseconds => stopwatch.ElapsedMilliseconds;
 
-        public void Begin(int attempt, int attemptSeed)
+        public void Begin(int attempt, long attemptSeed)
         {
             attemptNumber = attempt;
             seed = attemptSeed;
